feat: 채팅방 에러 처리 용 훅 작성
